Rename file to better reflect export name
